Handle shell pipe line input to zfec command.
zfec was throwing error when input was piped into it like below.

 `cat some-file | zfec -`

Now if we get illegal seek `IOError` and input is stdin we first read
entire content from stdin and calculate its length into fsize then pass
it to `filefec.encode_to_files` function.

If we get IOError and input is not stdin we raise an exception with
meaningful message. (But this should not occur).

Fixes: #3.