stop using PBR, unfortunately it isn't compatible with Versioneer
With PBR, setup.py is quite minimal, and most of the config was declared in
setup.cfg instead. Unfortunately PBR overrides the same hooks that Versioneer
uses to burn the correct version into a generated sdist/wheel. The
consequence was that installing the zfec-1.5.0 tarball resulted in pip
reporting zfec version "0+unknown" was installed. The self-reported ``from
zfec import __version__`` value was wrong too.

Removing PBR seems to be the easiest way to fix this. Modern setuptools can
do a lot of the declarative config, although with different names, so this is
mostly just a matter of rewriting setup.cfg.

I removed the following config, which I think is obsolete:

* setup_requires(stdeb), which was only used when running the "sdist_dsc"
  command (to build a debian package directly from our setup.py, which I
  believe Debian itself doesn't use)
* data_files: I think this was only used for stdeb/sdist_dsc
* test_suite: we now use Tox for all tests, so the responsibility for knowing
  how to run tests has moved to tox.ini
* extras_require (ed25519): I think this was added in error in 2013 (commit
  2a030713)
* zip_safe=False: This was zooko's preference, for easier debugging. It
  forces egg-based installs to unpack all contents (instead of keeping
  everything in a zip file and using import wizardry to find the right .py
  files). However most people use pip these days, which doesn't use eggs and
  unpacks everything during installation.

I added a bunch of files to MANIFEST.in because we no longer have access to
PBR's pull-files-from-git tools. I left out .gitattributes since I didn't
think it makes sense in the pypi distribution tarball, but I could be
convinced otherwise.

I moved the install_requires= into setup.py because I don't know how to make
a setup.cfg `[options.install_requires]` that includes the python2-only
qualifier on the "argparse" dependency.

closes #11