fec.c: switch back to old C syntax to appease windows compiler
The for-loop inline typedef was causing the appveyor C compiler to complain
about a syntax error. That compiler was also ignoring the -std=c99 argument,
which suggests it's pretty old. It might be a self-inflicted problem; our
misc/windows-build.cmd script might be pinning a specific compiler, which
could be old. Most of this windows tooling is cargo-culted from other
projects, so if anyone knows the Right Way To Do It, please please file a PR.