Automate the upload to PyPI via CI
Add job on the build workflow to automate the uploading to PyPi. Upload the
source distribution file (produced by sdist) and all built wheels (.whl files).

The implementation is based on the example from cibuildwheel, except for the tag
format. Expect version tags to be formatted as `zfec-x.y.z`.