avoid promoting `unsigned short` to `int` before comparing it
`n` is unsigned short.
`n - 1` is promoted to int because all unsigned short fit into int
(if they did not it would be promoted to unsigned int).

`row` is unsigned int.

`row < n - 1` is `(unsigned int) < (int)` is a compiler warning about
comparing types with different signedness.

`row + 1` involves no promotion because `row` is already an `unsigned int`.
`row + 1 < n` is `(unsigned int) < (unsigned short)` which is not a compiler
warning.

`row + 1 < n` is equivalent to `row < n - 1` because `row + 1 < n` is just
`row < n - 1` with 1 added to both sides.

`n - 1` was never a negative overflow because n is asserted to be greater than
or equal to 1 higher up in this function.

`row + 1` is never a positive overflow because row is a loop variable starting
from 0 and never exceeding n, which can be no more than SHORT_MAX.
SHORT_MAX + 1 is not a positive overflow for unsigned int.